backend/PythonService/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
requests==2.31.0
beautifulsoup4==4.12.2
cryptography==41.0.7
chromadb
sentence-transformers
nest-asyncio

# Document Intelligence
pdfplumber==0.10.3
PyPDF2==3.0.1
python-docx==1.1.0
pillow==10.1.0
pytesseract==0.3.10
